+++ NEWS.md
@@ -1,5 +1,6 @@
 # simqi 0.2.0
 
+- `vcov` arguments allow for custom variance-covariance matrices (like those you would have for "robust" standard errors).
 - Optimize `newdata` to spare user from having to create such a column in their prediction grid.
 - Documentation updates.